pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "land-audit"
version = "0.1.0"
description = "Orome Land Audit Engine - Wildlife management site analysis tool"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Orome", email = "info@orome.com"}
]
keywords = ["geospatial", "wildlife", "terrain-analysis", "camera-placement"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "py3dep>=0.16.0",
    "rasterio>=1.3.9",
    "numpy>=1.26.0",
    "shapely>=2.0.2",
    "geopandas>=0.14.0",
    "pyproj>=3.6.0",
    "scipy>=1.11.0",
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "click>=8.1.0",
    "tqdm>=4.66.0",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
    "scikit-learn>=1.3.0",
    "modal>=0.64.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[project.scripts]
land-audit = "land_audit.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=land_audit --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
